Renamed the hits components to results-list.
