+++ src/components/results-list/CRUKSearchkitResultsList.jsx
@@ -8,18 +8,18 @@ import {
 /**
  * Import our custom highlight accessor which won't mulch the array of objects.
  */
-import CRUKHighlightAccessor from './CRUKHighlightAccessor.jsx'
-import CRUKHitsAccessor from './CRUKHitsAccessor.jsx'
+import CRUKSearchkitResultsHighlightAccessor from './CRUKSearchkitResultsHighlightAccessor.jsx'
+import CRUKSearchkitResultsAccessor from './CRUKSearchkitResultsAccessor.jsx'
 /**
  * And the rest.
  */
-import CRUKHitsList from './../hitsList/HitsList.jsx';
+import CRUKSearchkitResultsListDisplay from './CRUKSearchkitResultsListDisplay.jsx';
 import CRUKSearchkitResult from './../result/CRUKSearchkitResult.jsx';
 
 /**
  * Define our class.
  */
-export default class CRUKSearchHits extends Hits {
+export default class CRUKSearchkitResultsList extends Hits {
 
   constructor(props) {
     super();
@@ -37,7 +37,7 @@ export default class CRUKSearchHits extends Hits {
     mod: 'search-results',
     hitsPerPage: 10,
     itemComponent: CRUKSearchkitResult,
-    listComponent: CRUKHitsList,
+    listComponent: CRUKSearchkitResultsListDisplay,
     scrollTo: true
   }
 
@@ -46,17 +46,18 @@ export default class CRUKSearchHits extends Hits {
      * Don't use super() here to avoid adding in both hits accessor and our
      * custom one.
      */
-    this.searchkit = this._getSearchkit()
-    if(this.searchkit){
-      this.accessor  = this.defineAccessor()
-      if(this.accessor){
+    this.searchkit = this._getSearchkit();
+
+    if (this.searchkit) {
+      this.accessor = this.defineAccessor();
+      if (this.accessor) {
         this.accessor = this.searchkit.addAccessor(this.accessor)
       }
       this.stateListenerUnsubscribe = this.searchkit.emitter.addListener(()=> {
         !this.unmounted && this.forceUpdate()
-      })
+      });
     } else {
-      console.warn("No searchkit found in props or context for " + this.constructor["name"])
+      console.warn('No searchkit found in props or context for ' + this.constructor['name']);
     }
 
     /**
@@ -64,26 +65,25 @@ export default class CRUKSearchHits extends Hits {
      * settings per highlight field.
      */
     if (this.props.CRUKHighlightFields) {
-      this.searchkit.addAccessor(
-        new CRUKHighlightAccessor(this.props.CRUKHighlightFields))
+      this.searchkit.addAccessor(new CRUKSearchkitResultsHighlightAccessor(this.props.CRUKHighlightFields));
     }
 
-    this.hitsAccessor = new CRUKHitsAccessor({ scrollTo: this.props.scrollTo })
-    this.searchkit.addAccessor(this.hitsAccessor)
+    this.hitsAccessor = new CRUKSearchkitResultsAccessor({ scrollTo: this.props.scrollTo });
+    this.searchkit.addAccessor(this.hitsAccessor);
   }
 
   render() {
-    let hits = this.getHits()
-    let hasHits = hits.length > 0
+    let hits = this.getHits();
+    let hasHits = hits.length > 0;
 
     if (!this.isInitialLoading() && hasHits) {
-      const {listComponent, mod, className, itemComponent, additionalFields} = this.props
+      const {listComponent, mod, className, itemComponent, additionalFields} = this.props;
+
       return renderComponent(listComponent, {
         hits, mod, className, itemComponent, additionalFields
-      })
+      });
     }
 
-    return null
-
+    return null;
   }
 }
